requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

django-debug-toolbar
django-filter
django-phonenumber-field[phonenumberslite]
django-rest-framework
djangorestframework-simplejwt
drf-nested-routers
pillow
stripe
